@@ -12,25 +12,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.0.2",
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/forms": "4.0.0",
-    "@angular/http": "4.0.0",
-    "@angular/material": "2.0.0-beta.3",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/router": "4.0.0",
+    "@angular/animations": "7.0.4",
+    "@angular/cdk": "^7.1.0",
+    "@angular/common": "7.0.4",
+    "@angular/compiler": "7.0.4",
+    "@angular/core": "7.0.4",
+    "@angular/forms": "7.0.4",
+    "@angular/http": "7.0.4",
+    "@angular/material": "7.1.0",
+    "@angular/platform-browser": "7.0.4",
+    "@angular/platform-browser-dynamic": "7.0.4",
+    "@angular/platform-server": "7.0.4",
+    "@angular/router": "7.0.4",
     "angular2-busy": "2.0.4",
     "core-js": "2.4.1",
     "hammerjs": "2.0.8",
-    "rxjs": "5.1.0",
-    "zone.js": "0.8.4"
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0",
-    "@angular/compiler-cli": "4.0.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "7.0.6",
+    "@angular/compiler-cli": "7.0.4",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "codelyzer": "~2.0.0",
@@ -39,12 +43,12 @@
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "0.2.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "0.2.2",
-    "karma-coverage-istanbul-reporter": "0.2.0",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "3.1.6"
   }
 }
